Refactor radio and text filter into once function
To reduce code duplication of filters and lebels. Combined the two search function into one, after a hours of googling, a simple (a or b) && c if statement to display the correct images for either just radio button filter or textfilter or both.

Also added a new label function reusing the code from previews version.